+++ src/array/codec/array_to_bytes/codec_chain.rs
@@ -359,7 +359,7 @@ impl ArrayCodecTraits for CodecChain {
         decoded_value: Vec<u8>,
         decoded_representation: &ArrayRepresentation,
     ) -> Result<Vec<u8>, CodecError> {
-        if decoded_value.len() != decoded_representation.size() {
+        if decoded_value.len() as u64 != decoded_representation.size() {
             return Err(CodecError::UnexpectedChunkDecodedSize(
                 decoded_value.len(),
                 decoded_representation.size(),
@@ -374,7 +374,7 @@ impl ArrayCodecTraits for CodecChain {
         decoded_value: Vec<u8>,
         decoded_representation: &ArrayRepresentation,
     ) -> Result<Vec<u8>, CodecError> {
-        if decoded_value.len() != decoded_representation.size() {
+        if decoded_value.len() as u64 != decoded_representation.size() {
             return Err(CodecError::UnexpectedChunkDecodedSize(
                 decoded_value.len(),
                 decoded_representation.size(),
